Bug-fix: the GPS time-of-week (in seconds) was kept as long, similarly to the UNIX time (in milliseconds)... therefore, it was losing the fractional part when converting from one to the other (this is an issue particularly with low-cost receivers, which often have non-integer GPS time-of-week)

+++ src/main/java/org/gogpsproject/Time.java
@@ -109,7 +109,7 @@ public class Time {
 	 *            (UNIX standard time in milliseconds)
 	 * @return GPS time in seconds
 	 */
-	private static long unixToGpsTime(long time) {
+	private static double unixToGpsTime(double time) {
 		// Shift from UNIX time (January 1, 1970 - msec)
 		// to GPS time (January 6, 1980 - sec)
 		time = time / Constants.MILLISEC_IN_SEC - Constants.UNIX_GPS_DAYS_DIFF * Constants.SEC_IN_DAY;
@@ -181,12 +181,12 @@ public class Time {
 		return tmpl;
 	}
 
-	public long getGpsTime(){
+	public double getGpsTime(){
 		return unixToGpsTime(msec);
 	}
 	
-	public long getRoundedGpsTime(){
-		long tow = unixToGpsTime((msec+499)/1000*1000);
+	public double getRoundedGpsTime(){
+		double tow = unixToGpsTime((msec+499)/1000*1000);
 		return tow;
 	}
 //
